Cleaned up test fixtures
